✨ feat(workflow): 添加获取工作流发布详情功能
- 实现DifyWorkflow类中的get_publish方法
- 添加应用ID为空时抛出异常的逻辑
- 更新位置类的坐标类型为浮点数

📝 docs(workflow_publish): 添加获取工作流发布详情示例

- 创建示例脚本展示如何使用get_publish方法
- 加载环境变量以便于API调用

✅ test(workflow): 增加获取工作流发布详情的单元测试

- 添加测试用例验证工作流发布详情的获取
- 验证空应用ID的异常抛出情况

+++ dify/app/workflow/schemas.py
@@ -6,8 +6,8 @@ from pydantic import BaseModel, Field
 class Position(BaseModel):
     """位置"""
 
-    x: Optional[int] = Field(default=None, description="x坐标")
-    y: Optional[int] = Field(default=None, description="y坐标")
+    x: Optional[float] = Field(default=None, description="x坐标")
+    y: Optional[float] = Field(default=None, description="y坐标")
 
 
 class WorkflowNodeData(BaseModel):
